:recycle: refactor : 북마크 기능 컴포넌트 분리
